[project @ 2003-01-22 23:36:42 by eburns] r=craigmcc
This checkin makes Facets work in the RI.  It is necessary because
FacesTag.findComponent() previously assumed that every UIComponent
instance had a non-null getParent().  This is not true of facets, who
are not their parent's children.  My fix is to introduce a new
"facetParent" component attribute, which is set when the facet is added
to the parent, and removed when the facet is removed.

Given that this attribute is set, we can use it in
FacesTag.findComponent() to traverse to the nearest NamingContainer.

M src/javax/faces/component/UIComponent.java

- Add attribute for facetParent, reference it in javadocs for addFacet(),
  removeFacet().

M src/javax/faces/component/UIComponentBase.java

- set/clear the FACET_PARENT attribute as a side effect of calling
  add/removeFacet().

M src/javax/faces/webapp/FacesTag.java

- Leverage the FACET_PARENT attribute to allow finding the nearest
  NamingContainer to a facet.

M test/javax/faces/component/UIComponentTestCase.java

- Test FACET_PARENT is set and cleared as specified.

Original author: eburns
Date: 2003-01-22 23:36:43+00:00