[project @ 2003-03-14 02:37:41 by craigmcc] Implement correct support for facets into the revised FacesTag.
In FacetTag, remove support for verifying a single child.  This should be
done at compile time (in a TLV) instead of at run time.

Refactor FacesTagTestCase to avoid tedious cut-n-paste copying of the
code that sets up the test tree of tags, so that additional tests can
be added easily.  Also, make sure that the tree is rendered twice in every
test (which simulates the postback scenario), and ensure that the output
text that is rendered (and the shape of the component tree) is identical
in every case.

Original author: craigmcc
Date: 2003-03-14 02:37:42+00:00