[project @ 2003-07-29 23:38:44 by horwat] [r=craigmcc]
In a non-jsp case the ViewHandler is responsible to make sure that the ResponseWriter is properly set.

Previously a concrete ServletResponseWriter implementation existed in the JSF-API but that has since been removed. Now in order to get the ResponseWriter, go through the default RenderKit and get the HtmlResponseWriter by passing in the HttpServletResponse writer and char encoding

Original author: horwat
Date: 2003-07-29 23:38:44+00:00