[project @ 2003-08-27 00:56:48 by craigmcc] Remodel the getState() and restoreState() methods to use a consistent, simple, style that will be the same as what we use in demo components (and therefore serve as a model of how component writers should do it). All unit tests pass -- the only changes to the tests were: - In UICommandBaseTestCase.testStateHolder(), modify the test for the   default ActionListener to reflect the fact that there are now two   default listeners. - In UIComponentBaseStateTestCase.testProcessState(), just throw   an exception back to the framework instead of swallowing it so you   can't see what happened.
In UICommandBase, implemented the pending save/restore of the "immediate"
flag, and adjusted the drop/add of the default action listener to pick the
correct one based on that flag.

In StateManager, had to modify StateHolderSaver so that it would not throw
an NPE when the "toSave" argument is null.  It would save an object create
to avoid this situation where StateHolderSaver instances are created.

Original author: craigmcc
Date: 2003-08-27 00:56:56+00:00