[project @ 2003-10-18 02:10:39 by craigmcc] [NEW] "rendered" property and phase processing
Changes to UIComponentBase:
* Add protected static isRendered(UIComponent component) that returns
  false if the component's own rendered property is false, or the
  component is nested in a parent component that has rendersChildren=true
  and rendered=false.
* Modify the processXxx() methods to not call the local component
  methods if the new static isRendered() method returns false for this
  component.

The net effect of these changes is to skip all processing in Apply Request
Values ... Render Response for components whose "rendered" property is false,
or who live inside a parent component (like UIPanel) whose "rendered"
property is false.

RI IMPACT:  Verify that tests still operate correctly.  Probably nothing
being tested in this regard.

DEMO IMPACT:  We should modify the tabbed panes demo to use "rendered"
as the mechanism by which a particular pane is to be displayed or not.
This will reduce the complexity of the Renderer considerably.

Original author: craigmcc
Date: 2003-10-18 02:10:40+00:00