[project @ 2004-01-16 18:15:19 by horwat] [r=visvan]
Bugtraq #4975242

If onclick is specified, added the "return false" statement. Also change the JavaScript event handler to "onclick" from "onmousedown".

The user still will not get the behavior they are expecting as described in the bug report. This is because they are expecting command_link to behave like output_link. They are confused since command_link looks like an "href" that it should behave like a standard link, instead it actually is a button. When pressed it behaves like a button and the page will be redisplayed and the cursor position will not be saved.

In the user's example as described in the bug report they should have used output_link instead. Also, their "onclick" attribute is a pass through attribute so the user themselves can specify whether "false" should be returned or not with that attribute.

Ed, we discussed this over the phone. The specification should be more clear about the command_link button-like behavior to avoid confusion like in this bug report.

Here are the modified files:

M src/com/sun/faces/renderkit/html_basic/CommandLinkRenderer.java
M systest/web/golden/taglib/command_link_test.txt
M web/test/RenderResponse_correct

Original author: horwat
Date: 2004-01-16 18:15:22+00:00