[project @ 2004-01-29 06:56:20 by craigmcc] [BugTraq 4983690] Fix Part III - iterator.remove() support additions
This is the third of three changes to UIComponentBase, dealing with
the Map returned by getFacets(), and includes robust unit tests.

TCK IMPACT:  There is a small tweak to the description of the Map
returned by getAttributes() that makes it match the actual implemented
behavior.  It is legal for an attribute to have a null value, but it is
still not legal to have a null key.

Original author: craigmcc
Date: 2004-01-29 06:56:22+00:00