[project @ 2008-01-16 03:09:49 by edburns] This is the first major JavaDoc change for JSF 2.0.
I am introducing a CSS style for making JSF 2.0 changes clearly
indicated in the JavaDocs to make it easy for reviewers to see what's
new.

These will be removed when the Proposed Final Draft version of the spec
is presented.

Roger, all spec changes going forward must be marked with this CSS
style.

For each each method, class, or member variable that is changed as a
result of work in JSF 2.0, use the css class "changed".  It's very
important that the CSS style appear in the method index of the JavaDoc.
Therefore, in the first paragraph of a javadoc entry, you need to add
the "changed" class to the <p> tag.  Like this:

     * <p class="changed">Return the singleton, stateless, thread-safe {@link
     * ResourceHandler} for this application.  The JSF implementation
     * must support the following techniques for declaring an alternate
     * implementation of <code>ResourceHandler</code>.</p>

For the rest of the javadoc entry, you need a separate div tag, after
the </p>, like this:

     * <div class="changed">

Then, you have to close out the </div> right before the @since 2.0, like this:

     * </div>

     * @since 2.0

You don't need to do anything to include the stylesheet that has the
"changed" class.  That's done for you by the ant task.

SECTION: API Changes

jsf-api/build.xml

- use the "stylesheet" attribute in the ant task for javadoc to include
  the jsf-api.css stylesheet.

- copy the images needed by the JSF 2.0 change bar styles.

jsf-api/doc/changed-new-in-2_0.png
jsf-api/doc/changed-new-in-2_0.psd

The binary artifacts for the change bars

jsf-api/src/jsf-api.css

The style sheet that includes the change bars.

jsf-api/src/javax/faces/application/Application.java

- The first file to use the "changed" style class.

Original author: edburns
Date: 2008-01-16 03:09:50+00:00