Issue: 280-EZComp MethodExpression passing
SECTION: Modified Files
----------------------------
M      jsf-api/src/javax/faces/webapp/pdl/PDLUtils.java

- Add new method retargetMethodExpressions()

Leverage the component metadata specified in section 4.3.2 for the
purpose of re-targeting any method expressions from the top level
component to the appropriate inner component. For each attribute that is
a MethodExpression (as indicated by the presence of a "method-signature"
attribute and the absence of a "type" attribute), the following action
must be taken:

Get the value of the applyTo attribute. If not found, log an error and
continue to the next attribute.

Find the inner component of the topLevelComponent with the id equal to
applyTo. For discussion, this component is called target. If not found,
log and error and continue to the next attribute.

For discussion the declared name of the attribute is called name.

In the attributes map of the topLevelComponent, look up the entry under
the key name. Assume the result is a ValueExpression. For discussion,
this is attributeValueExpression. If not found, log an error and
continue to the next attribute.

If name is equal to the string "action", or "actionListener" without the
quotes, assume target is an ActionSource2.

If name is equal to the string "validator", or "valueChangeListener"
without the quotes, assume target is an EditableValueHolder.

Call getExpressionString() on the attributeValueExpression and use that
string to create a MethodExpression of the appropriate signature for
name.

If name is not equal to any of the previously listed strings, call
getExpressionString() on the attributeValueExpression and use that
string to create a MethodExpression where the signature is created based
on the value of the "method-signature" attribute of the
<composite:attribute /> tag.

Let the resultant MethodExpression be called attributeMethodExpression
for discussion.

If name is equal to the string "action" without the quotes, call
ActionSource2.setActionExpression(javax.el.MethodExpression) on target,
passing attributeMethodExpression.

If name is equal to the string "actionListener" without the quotes, call
ActionSource.addActionListener(javax.faces.event.ActionListener) on
target, passing attributeMethodExpression wrapped in a
MethodExpressionActionListener.

If name is equal to the string "validator" without the quotes, call
EditableValueHolder.addValidator(javax.faces.validator.Validator) on
target, passing attributeMethodExpression wrapped in a
MethodExpressionValidator.

If name is equal to the string "valueChangeListener" without the quotes,
call
EditableValueHolder.addValueChangeListener(javax.faces.event.ValueChangeListener)
on target, passing attributeMethodExpression wrapped in a
MethodExpressionValueChangeListener.

Otherwise, look for a JavaBeans setter that matches name and assume it
takes a MethodExpression, passing attributeMethodExpression as that
expression. If such a setter does not exist, or does not take a
MethodExpression, log an error and continue to the next attribute.

M      jsf-ri/src/com/sun/faces/facelets/tag/composite/AttributeHandler.java

- Add type, method-signature, and applyTo attributes

M      jsf-ri/src/com/sun/faces/facelets/tag/jsf/CompositeComponentTagHandler.java

- Call the new retargetMethodExpressions() method

M      jsf-demo/ezcomp02/src/main/webapp/index.xhtml
M jsf-demo/ezcomp02/src/main/webapp/resources/ezcomp/loginPanel.xhtml

- Test new code