Issue: 265
SECTION: Modified Files
----------------------------
M      jsf-ri/src/com/sun/faces/config/processor/RenderKitConfigProcessor.java

- Modify to state that RenderKits can be decorated.

A      jsf-api/src/javax/faces/render/RenderKitWrapper.java

- Simple wrapper for RenderKit

SECTION: Spec Prose document Mods:

Add this to 11.4.5 Delegating Implementation Support

The implementation must also support decoration of a RenderKit
instance. At the point in time of when the <render-kit> element is
processed in an application configuration resources, if the current
RenderKitFactory already has a RenderKit instance for the
<render-kit-id> within the <render-kit> element, and the Class whose
fully qualified java class name is given as the value of the
<render-kit-class> element within the <render-kit> element has a
constructor that takes an RenderKit instance, the existing RenderKit for
that <render-kit-id> must be passed to that constructor, and the
RenderKit resulting from the executing of that constructor must be
passed to RenderKitFactory.addRenderKit().