Changes for issue 1197.
Introduce a new configuration option:

  - com.sun.faces.annotationScanPackages

This option allows the developer to dictate which packages in WEB-INF/classes and which jars/packages to scan in WEB-INF/lib.

Rules for the values:
   - values are whitespace delimited
   - to restrict packages in WEB-INF/classes, simply
      included which packages to scan (i.e. com.acme.package1 com.acme.package3)
  - to restrict which jars/packages are scanned the format of the
     entry is:
        * jar:<jar name>:<comma separated list of packages>
     
     So an example would be:  jar:a.jar:com.acme.package1,com.acme.package2

Note that anything not explicitly listed when using this option means it won't be scanned.  For instance, if a single package is listed to be scanned in WEB-INF/classes, no jar files in WEB-INF/lib will be scanned.

Putting it all together, an entry in the web.xml could look like:

<context-param>   
    <param-name>
        com.sun.faces.annotationScanPackages
    </param-name>
    <param-value>
         com.sun.faces.systest.model
         jar:a.jar:test2
    </param-value>
</context-param>

Only classes package com.sun.faces.systest.model and its sub-packages will be scanned for annotations.  Only classes
in package test2 and its sub-packages in a.jar will be scanned
for annotations.

Lastly, this configuration option is only relevant to the default AnnotationProvider implementation.  We'll need to document that.